+++ ydb/tools/mnc/lib/__init__.py
@@ -1,6 +1,7 @@
 from . import common
 from . import deploy_ctx
 from . import parted
+from . import progress
 from . import structure
 from . import templates
 from . import term
@@ -8,11 +9,14 @@ from . import tools
 from . import ydb_config
 from . import configs
 from . import init
+from . import service
+from . import deploy
 
 modules = [
     common,
     deploy_ctx,
     parted,
+    progress,
     structure,
     templates,
     term,
@@ -20,4 +24,6 @@ modules = [
     ydb_config,
     configs,
     init,
+    service,
+    deploy,
 ]
