--- ydb/tools/mnc/lib/legacy_commands/__init__.py
@@ -1,8 +1,0 @@
-from . import deploy
-from . import service
-
-
-modules = [
-    deploy,
-    service,
-]
